Bugfix Pt.2 & Chases trackingEntity (READ COMMIT DEC.)
Fixes a huge bug where the Warden randomly gets stuck:
This was caused by the "shouldContinue" method. If it returns true, then the Warden would try to pathfind to the SAME location it just moved to- ultimately making it pointless. This fixes that bug, so the pathfinding is ALWAYS refreshed.
The Warden also chases trackingEntity if it's nearby, suggested by BluePhoenix. It's a great idea, and definitely makes the Warden more challenging!