Warden Vibrations Work On Servers (READ DESC)
First thing's first, yes, I actually DID test on a server this time. No more week-long searches for the cause of a random crash caused by something I was oblivious of.

Second thing's second, it's a cleaner implementation of EntityPositionSource. EntityPositionSource was very obviously tampered with after they split Caves&Cliffs and didn't work properly, and I had to use a workaround that only worked in singleplayer. With this, it'll work properly both on clients AND servers, and matches what we would've gotten, had Caves&Cliffs never split and the Deep Dark/Warden never got delayed.