PLEASE READ DESC
So I might not be able to be on my pc as my mom has covid and she's in bed, the thing is my pc is in her room. And we need to make sure my dad is safe as he has a condition that makes his immune system REALLY weak, and he also has surgery for something else in a week.