Stinky Thief Mode (READ DESC)
If you want to play in Stinky Thief Mode, change the boolean in LivingEntityMixin to true. MAKE SURE THIS IS FALSE BEFORE COMMITING.

When STM is active, it'll effectively run a real-time world generator made entirely out of Sculk, which replaces all blocks.

If you want to add a stinky thief, copy the format in the tickMovement injection in LivingEntityMixin.