read desc
I'm updating one of my mods to allow pipes to transmit vibrations through them if something that can hear a GameEvent is on the other side of the pipeline. Might as well make it compatible with TWM, which is pretty much the only way this feature would have any use to begin with!