Warden Spawning Is Even Closer To 1.19 Snapshot (READ DESC)
Shriekers must be set off 4 times - should probably be set to 3 actually.

Wardens will whine and not emerge if emerge location is exposed to daylight and it isn't night.

WE NEED TO FIND A WAY TO SAVE THE AMOUNT OF SHRIEKS TO A PLAYER, THAT WILL SOLVE TONS OF PROBLEMS THAT WE CURRENTLY HAVE.